requirements.txt
requests>=2.25.1
progressbar2>=3.53.1